refactor(padding): 把resolveAllPadding方法抽成公共utils，重新处理venn的padding
